Make search results persist even after navigating away from the home page
