+++ src/components/Home.jsx
@@ -1,27 +1,19 @@
 import React, { useState, useEffect, Fragment } from "react";
-import Navbar from "./Navbar";
-// Header from "./components/Header";
 import Search from "./Search";
 import CardList from "./CardList";
-import useBookSearch from "../hooks/useBookSearch";
 import googleBookToAppBook from "../utils/googleBookToAppBook";
 
-export default function Home({ accessToken }) {
-  const [query, setQuery] = useState(undefined);
-  const [orderBy, setOrderBy] = useState("relevance");
-  const [pageNumber, setPageNumber] = useState(1);
-  const [areResultsLoading, setAreResultsLoading] = useState(false);
-
-  const { books, error, isLastPage, queryHistory } = useBookSearch(query, orderBy, pageNumber, setAreResultsLoading); // 'books' has search results
-
-  const handleSubmit = (e, searchTerm, orderBy) => {
-    e.preventDefault();
-    setQuery(searchTerm);
-    setOrderBy(orderBy);
-    setPageNumber(1);
-    // setModal(false)
-  };
-
+export default function Home({
+  query,
+  accessToken,
+  books,
+  error,
+  isLastPage,
+  queryHistory,
+  areResultsLoading,
+  handleSubmit,
+  setPageNumber,
+}) {
   useEffect(() => {
     const cardListElement = document.getElementById("cardList");
     cardListElement.addEventListener("scroll", handleScroll);
@@ -38,7 +30,6 @@ export default function Home({ accessToken }) {
   return (
     <Fragment>
       <Search handleSubmit={handleSubmit} error={error} queryHistory={queryHistory} />
-      {/* <CardList books={books.map(googleBookToAppBook)} isLastPage={isLastPage}/> */}
       <CardList
         books={books.map(googleBookToAppBook)}
         isLastPage={isLastPage}
@@ -46,7 +37,6 @@ export default function Home({ accessToken }) {
         accessToken={accessToken}
         buttonType="favorite"
       />
-      {/* <div>{areResultsLoading ? "Loading" : ""}</div> */}
       {areResultsLoading === true && <div className="loading-msg">Loading...</div>}
     </Fragment>
   );
